requirements.txt
Django==4.2.4
django-cors-headers==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.0.0
python-dotenv
